backend/auth_api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic>=2.7.4,<3.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.22
pydantic-settings>=2.4.0,<3.0.0
email-validator==2.1.0.post1
python-dotenv==1.0.0
